Artificial Assistant with facial Recognition
It is a system that has been developed to perform defined task for user.
It reduces unnecessary need of extra manpower to perform a task that can be implemented
digitally. The system takes special care of security by checking the Authentication at very
beginning. It has the ability to recognize the user using face recognition. The recognized
user can give defined voice command which the system takes as input. The tasks like Web
Queries, Mail Handling, Reminder Setting, Simple arithmetic calculations, Weather
Checking can be performed by voice command and the result or output is also given in
voice form.